backend/requirements.txt
Flask==3.1.1
pydantic==2.11.4
python-dotenv==1.1.0
flask-cors==5.0.1